auth_service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pyjwt==2.8.0
python-dotenv==1.0.0
bcrypt==4.0.1
pytest==7.4.2
pytest-cov==4.1.0
azure-cosmos==4.5.1
azure-storage-blob==12.17.0
